% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sd_1cmt_linear_infusion.R
\name{calc_sd_1cmt_linear_infusion}
\alias{calc_sd_1cmt_linear_infusion}
\title{Calculate C(t) for a 1-compartment linear model after a single infusion}
\usage{
calc_sd_1cmt_linear_infusion(t, CL, V, dose, tinf)
}
\arguments{
\item{t}{Time after dose (h)}

\item{CL}{Clearance (L/h)}

\item{V}{Central volume of distribution (L)}

\item{dose}{Dose}

\item{tinf}{Duration of infusion (h)}
}
\value{
Concentration of drug at requested time (\code{t}) after a single dose, given provided set of parameters and variables.
}
\description{
Calculate C(t) for a 1-compartment linear model after a single infusion
}
\examples{
Ctrough <- calc_sd_1cmt_linear_infusion(t=0:24, CL=6, V=25, dose=600, tinf=1)

}
\references{
Bertrand J & Mentre F (2008). Mathematical Expressions of the Pharmacokinetic and Pharmacodynamic Models
implemented in the Monolix software. \url{http://lixoft.com/wp-content/uploads/2016/03/PKPDlibrary.pdf}
}

