% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sd_1cmt_linear_oral_0.R
\name{calc_sd_1cmt_linear_oral_0}
\alias{calc_sd_1cmt_linear_oral_0}
\title{Calculate C(t) for a 1-compartment linear model with zero-order absorption after a single oral dose}
\usage{
calc_sd_1cmt_linear_oral_0(t, CL, V, dur, dose)
}
\arguments{
\item{t}{Time after dose (h)}

\item{CL}{Clearance (L/h)}

\item{V}{Central volume of distribution (L)}

\item{dur}{Duration of zero-order absorption (h)}

\item{dose}{Dose}
}
\value{
Concentration of drug at requested time (\code{t}) after a single dose, given provided set of parameters and variables.
}
\description{
Calculate C(t) for a 1-compartment linear model with zero-order absorption after a single oral dose
}
\examples{
Ctrough <- calc_sd_1cmt_linear_oral_0(t=0:24, CL=6, V=25, dur=1.5, dose=600)

}
\references{
Bertrand J & Mentre F (2008). Mathematical Expressions of the Pharmacokinetic and Pharmacodynamic Models
implemented in the Monolix software. \url{http://lixoft.com/wp-content/uploads/2016/03/PKPDlibrary.pdf}
}
\author{
Justin Wilkins, \email{justin.wilkins@occams.com}
}
