% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sd_1cmt_linear_oral_1.R
\name{calc_sd_1cmt_linear_oral_1}
\alias{calc_sd_1cmt_linear_oral_1}
\title{Calculate C(t) for a 1-compartment linear model with first-order absorption after a single oral dose}
\usage{
calc_sd_1cmt_linear_oral_1(t, CL, V, ka, dose)
}
\arguments{
\item{t}{Time after dose (h)}

\item{CL}{Clearance (L/h)}

\item{V}{Central volume of distribution (L)}

\item{ka}{First order absorption rate constant (/h)}

\item{dose}{Dose}
}
\value{
Concentration of drug at requested time (\code{t}) after a single dose, given provided set of parameters and variables.
}
\description{
Calculate C(t) for a 1-compartment linear model with first-order absorption after a single oral dose
}
\examples{
Ctrough <- calc_sd_1cmt_linear_oral_1(t=0:24, CL=6, V=25, ka=1.1, dose=600)

}
\references{
Bertrand J & Mentre F (2008). Mathematical Expressions of the Pharmacokinetic and Pharmacodynamic Models
implemented in the Monolix software. \url{http://lixoft.com/wp-content/uploads/2016/03/PKPDlibrary.pdf}
}

