% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sd_2cmt_linear_infusion.R
\name{calc_sd_2cmt_linear_infusion}
\alias{calc_sd_2cmt_linear_infusion}
\title{Calculate C(t) for a 2-compartment linear model after a single infusion}
\usage{
calc_sd_2cmt_linear_infusion(t, CL, V1, V2, Q, dose, tinf)
}
\arguments{
\item{t}{Time after dose (h)}

\item{CL}{Clearance (L/h)}

\item{V1}{Central volume of distribution (L)}

\item{V2}{Peripheral volume of distribution (L)}

\item{Q}{Intercompartmental clearance (L/h)}

\item{dose}{Steady state dose}

\item{tinf}{Duration of infusion (h)}
}
\value{
Concentration of drug at requested time (\code{t}) after a single dose, given provided set of parameters and variables.
}
\description{
Calculate C(t) for a 2-compartment linear model after a single infusion
}
\examples{
Ctrough <- calc_sd_2cmt_linear_infusion(t = 11.75, CL = 7.5, V1 = 20, V2 = 30, Q = 0.5,
    dose = 10, tinf = 1)

}
\references{
Bertrand J & Mentre F (2008). Mathematical Expressions of the Pharmacokinetic and Pharmacodynamic Models
implemented in the Monolix software. \url{http://lixoft.com/wp-content/uploads/2016/03/PKPDlibrary.pdf}

Rowland M, Tozer TN. Clinical Pharmacokinetics and Pharmacodynamics: Concepts and Applications (4th). Lippincott Williams & Wilkins, Philadelphia, 2010.
}
\author{
Justin Wilkins, \email{justin.wilkins@occams.com}
}
