% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ss_1cmt_linear_oral_1_lag.R
\name{calc_ss_1cmt_linear_oral_1_lag}
\alias{calc_ss_1cmt_linear_oral_1_lag}
\title{Calculate C(t) for a 1-compartment linear model with first-order oral absorption at steady state, with lag time}
\usage{
calc_ss_1cmt_linear_oral_1_lag(tad, CL, V, ka, dose, tlag, tau)
}
\arguments{
\item{tad}{Time after dose (h)}

\item{CL}{Clearance (L/h)}

\item{V}{Central volume of distribution (L)}

\item{ka}{First order absorption rate constant (/h)}

\item{dose}{Steady state dose}

\item{tlag}{Lag time (h)}

\item{tau}{Dosing interval (h)}
}
\value{
Concentration of drug at requested time (\code{tad}) after dose, given provided set of parameters and variables.
}
\description{
Calculate C(t) for a 1-compartment linear model with first-order oral absorption at steady state, with lag time
}
\examples{
Ctrough <- calc_ss_1cmt_linear_oral_1_lag(tad=0:36, CL=2, V=25, dose=600,
    ka=0.25, tlag=0.75, tau=24)

}
\references{
Bertrand J & Mentre F (2008). Mathematical Expressions of the Pharmacokinetic and Pharmacodynamic Models
implemented in the Monolix software. \url{http://lixoft.com/wp-content/uploads/2016/03/PKPDlibrary.pdf}
}
\author{
Justin Wilkins, \email{justin.wilkins@occams.com}
}
