% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ss_2cmt_linear_bolus.R
\name{calc_ss_2cmt_linear_bolus}
\alias{calc_ss_2cmt_linear_bolus}
\title{Calculate C(t) for a 2-compartment linear model with IV bolus dosing at steady-state}
\usage{
calc_ss_2cmt_linear_bolus(tad, CL, V1, V2, Q, dose, tau)
}
\arguments{
\item{tad}{Time after last dose (h)}

\item{CL}{Clearance (L/h)}

\item{V1}{Central volume of distribution (L)}

\item{V2}{Peripheral volume of distribution (L)}

\item{Q}{Intercompartmental clearance (L/h)}

\item{dose}{Steady state dose}

\item{tau}{Dosing interval (h)}
}
\value{
Concentration of drug at requested time after last dose (\code{tad}) at steady state, given regular IV bolus dosing and provided set of parameters and variables.
}
\description{
Calculate C(t) for a 2-compartment linear model with IV bolus dosing at steady-state
}
\examples{
Ctrough <- calc_ss_2cmt_linear_bolus(tad = 11.75, CL = 7.5, V1 = 20, V2 = 30, Q = 0.5,
    dose = 10, tau=24)

}
\references{
Bertrand J & Mentre F (2008). Mathematical Expressions of the Pharmacokinetic and Pharmacodynamic Models
implemented in the Monolix software. \url{http://lixoft.com/wp-content/uploads/2016/03/PKPDlibrary.pdf}

Rowland M, Tozer TN. Clinical Pharmacokinetics and Pharmacodynamics: Concepts and Applications (4th). Lippincott Williams & Wilkins, Philadelphia, 2010.
}
\author{
Justin Wilkins, \email{justin.wilkins@occams.com}
}
