% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_na.R
\name{count_na}
\alias{count_na}
\title{Count the number of NA values in a vector.}
\usage{
count_na(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
An integer containing the number of NA values in the input vector.
}
\description{
Count the number of NA values in a vector.
}
\examples{
\dontrun{
 count_na(c(0,5,7,NA,3,3,NA))
}

}
\author{
Justin Wilkins, \email{justin.wilkins@occams.com}
}
