% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_signif.R
\name{fmt_signif}
\alias{fmt_signif}
\title{Format a number with the correct number of significant digits and trailing zeroes.}
\usage{
fmt_signif(x, digits = 3)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{digits}{The number of significant digits values should have (defaults to 3).}
}
\value{
A string containing the properly-formatted number.
}
\description{
Format a number with the correct number of significant digits and trailing zeroes.
}
\examples{
\dontrun{
 fmt_signif(c(36.44, 0.0002, 3336.7), digits=3)
}

}
\author{
Justin Wilkins, \email{justin.wilkins@occams.com}
}
