% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.R
\name{geomean}
\alias{geomean}
\title{Calculate a geometric mean.}
\usage{
geomean(x, na.rm = F, neg.rm = F)
}
\arguments{
\item{x}{A vector.}

\item{na.rm}{Flag for removing \code{NA} values (defaults to \code{FALSE}).}

\item{neg.rm}{Flag for removing negative or zero values (defaults to \code{FALSE}).}
}
\value{
The geometric mean of the input vector. If \code{neg.rm} is \code{FALSE} and values <= 0 are present, \code{NA} will be returned.
}
\description{
Calculate a geometric mean.
}
\examples{
\dontrun{
 geomean(myvector)
}

}
\author{
Justin Wilkins, \email{justin.wilkins@occams.com}
}
