% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probinfo.r
\name{get_probinfo}
\alias{get_probinfo}
\title{Extract problem and estimation information from a NONMEM output object.}
\usage{
get_probinfo(x, sigdig = 6, est.step = NULL)
}
\arguments{
\item{x}{A NONMEM output object generated using \code{\link{read_nm}}.}

\item{sigdig}{Specifies the number of significant digits to be provided (default=6).}

\item{est.step}{Specifies which estimation step to return parameters from (default is the last).}
}
\value{
A list with elements \code{prob_title}, \code{prog_info},
\code{lic_info}, \code{prob_times} (data frame of start, stop and elapsed
times), \code{prob_info} (data frame of data set and model dimensions) and
\code{est_info} (data frame of estimation results for \code{est.step}).
}
\description{
Extract problem and estimation information from a NONMEM output object.
}
\examples{
\dontrun{
 nmOutput <- read_nm("run315.xml")
 probInfo <- get_probinfo(nmOutput)
}

}
\seealso{
NONMEM (\url{https://www.iconplc.com/solutions/technologies/nonmem})
}
