% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shrinkage.R
\name{get_shrinkage}
\alias{get_shrinkage}
\title{Extract shrinkage estimates from a NONMEM output object.}
\usage{
get_shrinkage(x, output = "eta", sigdig = 3)
}
\arguments{
\item{x}{A NONMEM output object generated using \code{\link{read_nm}}.}

\item{output}{A flag specifying the shrinkage estimates to be output. Valid flag
values are \code{eta} (the default), \code{epsilon}, or \code{all}.}

\item{sigdig}{Specifies the number of significant digits to be provided (default=3).}
}
\value{
A named vector of NONMEM shrinkage estimates, or in the case of \code{all},
a list of named vectors.

\code{eta} returns a vector of ETA shrinkages, as reported by NONMEM.
\code{epsilon} returns EPSILON shrinkage, as reported by NONMEM.
\code{all} returns both ETA and EPSILON shrinkage estimates as a list of vectors.
}
\description{
Extract shrinkage estimates from a NONMEM output object.
}
\examples{
\dontrun{
 nmOutput <- read_nm("run315.xml")
 shr <- get_shrinkage(nmOutput, output="all")
}

}

