% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOmega.R
\name{getOmega}
\alias{getOmega}
\title{Extract structural model parameter estimates from a NONMEM output object.}
\usage{
getOmega(x, output = "est")
}
\arguments{
\item{x}{A NONMEM output object generated using \code{\link{readNM}}.}

\item{output}{A flag specifying the matrix or matrices to be output. Valid flag values are "est" (the default),
"se", "rse", "cor", "cse", or "all".}
}
\value{
A symmetrical matrix, or a list of symmetrical matrices if \code{all} is specified.

\code{est} returns the estimated OMEGA variance-covariance matrix.
\code{se} returns the standard errors for the estimated OMEGA variance-covariance matrix.
\code{rse} returns the relative standard errors for the estimated OMEGA variance-covariance matrix (se/est*100).
\code{cor} returns the correlation matrix matrix.
\code{cse} returns the standard errors for the correlation matrix.
\code{all} returns all available OMEGA matrices.
}
\description{
Extract structural model parameter estimates from a NONMEM output object.
}
\examples{
 nmOutput <- readNM("run315.xml")
 omegas   <- getOmega(nmOutput)
 omegaRSEs <- getOmega(nmOutput, "rse")
}

