% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTheta.R
\name{getTheta}
\alias{getTheta}
\title{Extract structural model parameter estimates from a NONMEM output object.}
\usage{
getTheta(x)
}
\arguments{
\item{x}{A NONMEM output object generated using \code{\link{readNM}}.}
}
\value{
A named vector of NONMEM model parameter estimates.
}
\description{
Extract structural model parameter estimates from a NONMEM output object.
}
\examples{
 nmOutput <- readNM("run315.xml")
 thetas <- getTheta(nmOutput)
}

