% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{gm}
\alias{gm}
\title{Calculate geometric mean}
\usage{
gm(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
The geometric mean. \code{NA} is returned if there are any non-positive elements in \code{x}.
}
\description{
Calculate geometric mean
}
\examples{
gm(c(0.5, 7, 8, 5))

}
\author{
Justin Wilkins, \email{justin.wilkins@occams.com}
}
