% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pcv}
\alias{pcv}
\title{Calculate percentage coefficient of variation}
\usage{
pcv(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{na.rm}{A logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
The percentage coefficient of variation.
}
\description{
Calculate percentage coefficient of variation
}
\examples{
pcv(rnorm(50, 5, 7.56))

}
\author{
Justin Wilkins, \email{justin.wilkins@occams.com}
}
