% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nmtables.R
\name{read_nm_std_ext}
\alias{read_nm_std_ext}
\title{Read a standard NONMEM extension file}
\usage{
read_nm_std_ext(fileName, extension, directory = NULL, ...)
}
\arguments{
\item{fileName}{The filename (with directory name, if applicable) to read
(with or without the extension)}

\item{extension}{The file extension to optionally append (preferably starting
with a ".")}

\item{directory}{The directory to look for files within.  If NULL, uses the
current directory.}

\item{...}{Passed to \code{read_nm_multi_table()}}
}
\value{
NULL if the file does not exist or the value of
  \code{read_nm_multi_table()} if it does exist.
}
\description{
Read a standard NONMEM extension file
}
\examples{
\dontrun{
read_nm_std_ext("run1", "phi")
}
}
