% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNM.R
\name{readNM}
\alias{readNM}
\title{Read NONMEM 7.2+ output into a list of lists.}
\usage{
readNM(fileName)
}
\arguments{
\item{fileName}{A NONMEM XML output file (e.g. "run315.xml").}
}
\value{
A list of lists corresponding to the NONMEM output object.
}
\description{
Read NONMEM 7.2+ output into a list of lists.
}
\examples{
nmOutput <- readNM("run315.xml")
}

