% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNMcov.R
\name{readNMcov}
\alias{readNMcov}
\title{Read in the NONMEM variance-covariance matrix.}
\usage{
readNMcov(fileName)
}
\arguments{
\item{fileName}{Root filename for the NONMEM run (e.g. "run315").

This function reads the ".cov" NONMEM output table, and will return an error if this
is missing.}
}
\value{
A symmetrical variance-covariance matrix covering all model parameters.
}
\description{
Read in the NONMEM variance-covariance matrix.
}
\examples{
 nmVcov <- readNMcov("run315")
}

