% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulation.R
\name{run_simulation}
\alias{run_simulation}
\title{Provide concentration-time simulations}
\usage{
run_simulation(model = "1cmt_oral", doseTimes, obsTimes, doses, params,
  summarize = F)
}
\arguments{
\item{summarize}{}
}
\description{
Provide concentration-time simulations
}

