% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleOmega.R
\name{sampleOmega}
\alias{sampleOmega}
\title{Sample from the multivariate normal distribution using the OMEGA variance-covariance matrix to generate new sets of simulated ETAs from NONMEM output.}
\usage{
sampleOmega(nmRun, n)
}
\arguments{
\item{nmRun}{Root filename for the NONMEM run (e.g. "run315").}

\item{n}{Number of samples required.}
}
\value{
A data frame containing \code{n} samples from the multivariate normal distribution, using
the estimated NONMEM OMEGA variance-covariance matrix. This provides \code{n} sets of ETA estimates
suitable for simulation of new patients.
}
\description{
Sample from the multivariate normal distribution using the OMEGA variance-covariance matrix to generate new sets of simulated ETAs from NONMEM output.
}
\examples{
 omDist <- sampleOmega("run315", 5000)
}

