% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSigma.R
\name{sampleSigma}
\alias{sampleSigma}
\title{Sample from the multivariate normal distribution using the SIGMA variance-covariance matrix to generate new sets of simulated EPSILONs from NONMEM output.}
\usage{
sampleSigma(nmRun, n)
}
\arguments{
\item{nmRun}{Root filename for the NONMEM run (e.g. "run315").}

\item{n}{Number of samples required.}
}
\value{
A data frame containing \code{n} samples from the multivariate normal distribution, using
the estimated NONMEM SIGMA variance-covariance matrix. This provides \code{n} sets of EPSILON estimates
suitable for simulation of new datasets.
}
\description{
Sample from the multivariate normal distribution using the SIGMA variance-covariance matrix to generate new sets of simulated EPSILONs from NONMEM output.
}
\examples{
 sigDist <- sampleSigma("run315", 5000)
}

