% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vplayout.R
\name{vplayout}
\alias{vplayout}
\title{Helper function for arranging plots on a page.}
\usage{
vplayout(x, y)
}
\arguments{
\item{x}{Row.}

\item{y}{Column.}
}
\value{
A layout for the viewport.
}
\description{
Helper function for arranging plots on a page.
}
\examples{
library(ggplot2)
library(grid)
df <- data.frame(gp = factor(rep(letters[1:3], each = 10)), y = rnorm(30))
plot1 <- ggplot(df, aes(gp, y)) + geom_point()
plot2 <- ggplot(df, aes(y, gp)) + geom_point()

grid.newpage()
pushViewport(viewport(layout = grid.layout(1, 2)))
print(plot1, vp = vplayout(1, 1))
print(plot2, vp = vplayout(1, 2))

}
\seealso{
\code{\link[grid]{grid.newpage}}, \code{\link[grid]{viewport}}, \code{\link[grid]{pushViewport}}, \code{\link{print}}.
}

