% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_probinfo.r
\name{get_probinfo}
\alias{get_probinfo}
\title{Extract problem information from a NONMEM output object.}
\usage{
get_probinfo(x)
}
\arguments{
\item{x}{A NONMEM output object generated using \code{\link{read_nm}}.

#' @examples
\dontrun{
 nmOutput <- read_nm("run315.xml")
 probInfo <- get_probinfo(nmOutput)
}}
}
\description{
Extract problem information from a NONMEM output object.
}
