% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_blq.R
\name{itrans_blq_linear}
\alias{itrans_blq_linear}
\alias{itrans_blq_log}
\title{Inverse transformation for linear BLQ data}
\usage{
itrans_blq_linear(lloq)

itrans_blq_log(lloq, base)
}
\arguments{
\item{lloq}{The value of the lower limit of quantification as a numeric
scalar}

\item{base}{The base for the logarithm}
}
\value{
A function of \code{x} that replaces \code{x < lloq} with \code{lloq}
}
\description{
For ggplot2 scales.
}
\section{Functions}{
\itemize{
\item \code{itrans_blq_log}: Log-scale inverse transform
}}

\seealso{
Other BLQ Transformation: 
\code{\link{blq_trans}()},
\code{\link{breaks_blq_general}()},
\code{\link{estimate_lloq}()},
\code{\link{ftrans_blq_linear}()},
\code{\link{label_blq}()}
}
\concept{BLQ Transformation}
