% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_blq.R
\name{label_blq}
\alias{label_blq}
\title{Label axes with censoring labels for BLQ}
\usage{
label_blq(lloq, lloq_text)
}
\arguments{
\item{lloq}{The value of the lower limit of quantification as a numeric
scalar}

\item{lloq_text}{The text to use on the axis to indicate values \code{<
lloq}.  It will be automatically set to \code{paste0("<", lloq)} if
missing.}
}
\value{
A function of \code{x} which returns the formatted values.
}
\description{
For ggplot2 scales.
}
\seealso{
Other BLQ Transformation: 
\code{\link{blq_trans}()},
\code{\link{breaks_blq_general}()},
\code{\link{estimate_lloq}()},
\code{\link{ftrans_blq_linear}()},
\code{\link{itrans_blq_linear}()}
}
\concept{BLQ Transformation}
