% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmxTools-package.R
\docType{package}
\name{pmxTools-package}
\alias{pmxTools}
\alias{pmxTools-package}
\title{pmxTools: Pharmacometric and Pharmacokinetic Toolkit}
\description{
Pharmacometric tools for common data analytical tasks; closed-form solutions for calculating concentrations at given times after dosing based on compartmental PK models (1-compartment, 2-compartment and 3-compartment, covering infusions, zero- and first-order absorption, and lag times, after single doses and at steady state, per Bertrand & Mentre (2008) \url{https://www.facm.ucl.ac.be/cooperation/Vietnam/WBI-Vietnam-October-2011/Modelling/Monolix32_PKPD_library.pdf}); parametric simulation from NONMEM-generated parameter estimates and other output; and parsing, tabulating and plotting results generated by Perl-speaks-NONMEM (PsN).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kestrel99/pmxTools}
  \item \url{https://kestrel99.github.io/pmxTools/}
  \item Report bugs at \url{https://github.com/kestrel99/pmxTools/issues}
}

}
\author{
\strong{Maintainer}: Justin Wilkins \email{justin.wilkins@occams.com} (\href{https://orcid.org/0000-0002-7099-9396}{ORCID})

Authors:
\itemize{
  \item Bill Denney \email{wdenney@humanpredictions.com} (\href{https://orcid.org/0000-0002-5759-428X}{ORCID})
  \item Rik Schoemaker \email{rik.schoemaker@occams.com}
}

Other contributors:
\itemize{
  \item Satyaprakash Nayak \email{sn248@cornell.edu} [contributor]
  \item Leonid Gibiansky \email{lgibiansky@quantpharm.com} [contributor]
  \item Andrew Hooker \email{Andrew.Hooker@farmbio.uu.se} [contributor]
  \item E. Niclas Jonsson \email{niclas.jonsson@pharmetheus.com} [contributor]
  \item Mats O. Karlsson \email{mats.karlsson@farmbio.uu.se} [contributor]
  \item John Johnson \email{johndjohnson@gmail.com} [contributor]
}

}
\keyword{internal}
