% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nm.R
\name{read_nm}
\alias{read_nm}
\title{Read NONMEM 7.2+ output into a list of lists.}
\usage{
read_nm(fileName)
}
\arguments{
\item{fileName}{A NONMEM XML output file (e.g. "run315.xml").}
}
\value{
A list of lists corresponding to a NONMEM output object.
}
\description{
Read NONMEM 7.2+ output into a list of lists.
}
\examples{
\dontrun{
nmOutput <- read_nm("run315.xml")
}

}
\seealso{
NONMEM (\url{http://www.iconplc.com/innovation/nonmem/})
}
\author{
Justin Wilkins, \email{justin.wilkins@occams.com}
}
