% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_scm.R
\name{read_scm}
\alias{read_scm}
\title{Read PsN SCM output into a format suitable for further use.}
\usage{
read_scm(dir, startPhase = "forward")
}
\arguments{
\item{dir}{A PsN SCM folder (containing \code{scmlog.txt}).}

\item{startPhase}{Where to start collating the output; can be \code{"forward"} (the default) or \code{"backward"}.}
}
\value{
A list of data frames, containing
  \item{forward}{all models evaluated during the forward inclusion step of
  covariate model building}
  \item{forwardSummary}{the covariate relationships selected at each forward
  step}
  \item{backward}{all models evaluated during the backward elimination step of covariate
  model building}
  \item{backwardSummary}{the covariate relationships eliminated at each backward step}
}
\description{
\code{read_scm} returns a summary of a PsN SCM (stepwise covariate modeling)
procedure. It depends on the presence of an \code{scmlog.txt} file in the
specified directory.
}
\examples{
\dontrun{
scm <- readSCM("E:/DrugX/ModelDevelopment/scm310")
}

}

